# pokeRest/go_gin_autogen/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    name: SYM
    url: https://twitter.com/SYM_souten
  description: PokeRest API
  title: PokeRest
  version: "1.0"
servers:
- url: http://localhost:3000/PokeRest
...
paths:
...
  /moves/{moveId}:
    get:
      description: 技取得
      operationId: get-Moves-Id
      parameters:
      - description: 技ID
        explode: false
        in: path
        name: moveId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Move'
          description: OK
      summary: GET Move
      tags:
      - Pokemon Base Data
...
  /pokemons/moves/{pokemonId}:
    get:
      description: 指定ポケモンの技一覧取得
      operationId: get-pokemon-moves-pokemonId
      parameters:
      - description: ポケモンID
        explode: false
        in: path
        name: pokemonId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PokemonMoves'
          description: OK
      summary: GET Moves of Specify Pokemon
      tags:
      - Pokemon Base Data
...
components:
...
  schemas:
...
    Move:
      example:
        pp: 5.962133916683182
        isCanGuard: true
        species: species
        specialMovePower: 6.027456183070403
        isContactMove: true
        name: name
        accuracy: 1.4658129805029452
        id: id
        power: 0.8008281904610115
        type: type
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        species:
          description: 分類（物理/特殊）
          type: string
        power:
          description: 威力
          type: number
        specialMovePower:
          description: 威力（ダイマックス技等）
          type: number
        accuracy:
          description: 命中
          type: number
        pp:
          description: PP
          type: number
        isContactMove:
          description: 接触/非接触
          type: boolean
        isCanGuard:
          description: 守る
          type: boolean
      required:
      - accuracy
      - id
      - isCanGuard
      - isContactMove
      - name
      - power
      - pp
      - specialMovePower
      - type
      type: object
...
    Moves:
      example:
        moves:
        - pp: 5.962133916683182
          isCanGuard: true
          species: species
          specialMovePower: 6.027456183070403
          isContactMove: true
          name: name
          accuracy: 1.4658129805029452
          id: id
          power: 0.8008281904610115
          type: type
        - pp: 5.962133916683182
          isCanGuard: true
          species: species
          specialMovePower: 6.027456183070403
          isContactMove: true
          name: name
          accuracy: 1.4658129805029452
          id: id
          power: 0.8008281904610115
          type: type
        count: 5.637376656633329
      properties:
        moves:
          items:
            $ref: '#/components/schemas/Move'
          type: array
        count:
          type: number
      required:
      - count
      - moves
      type: object
...
    PokemonMoves:
      description: 指定ポケモンが覚える技一覧
      example:
        moves:
        - moves:
          - pp: 5.962133916683182
            isCanGuard: true
            species: species
            specialMovePower: 6.027456183070403
            isContactMove: true
            name: name
            accuracy: 1.4658129805029452
            id: id
            power: 0.8008281904610115
            type: type
          - pp: 5.962133916683182
            isCanGuard: true
            species: species
            specialMovePower: 6.027456183070403
            isContactMove: true
            name: name
            accuracy: 1.4658129805029452
            id: id
            power: 0.8008281904610115
            type: type
          count: 5.637376656633329
        - moves:
          - pp: 5.962133916683182
            isCanGuard: true
            species: species
            specialMovePower: 6.027456183070403
            isContactMove: true
            name: name
            accuracy: 1.4658129805029452
            id: id
            power: 0.8008281904610115
            type: type
          - pp: 5.962133916683182
            isCanGuard: true
            species: species
            specialMovePower: 6.027456183070403
            isContactMove: true
            name: name
            accuracy: 1.4658129805029452
            id: id
            power: 0.8008281904610115
            type: type
          count: 5.637376656633329
        id: id
      properties:
        id:
          type: string
        moves:
          items:
            $ref: '#/components/schemas/Moves'
          type: array
      required:
      - id
      - moves
      type: object
...